move everything to diagnostics channel
